Create captive portal util, misc fixes
