+++ src/api/controllers/auth/login.js
@@ -7,9 +7,7 @@ const log = require('../../utils/log.js')(module);
 const errorHandler = require('../../utils/errorHandler');
 const hasCartPaid = require('../../utils/hasCartPaid');
 const validateBody = require('../../middlewares/validateBody');
-const redis = require('../../utils/redis');
-const getIP = require('../../utils/getIP');
-const hasOrgaPermission = require('../../utils/hasOrgaPermission');
+const captivePortal = require('../../utils/captivePortal');
 
 const CheckLogin = [
   check('username').exists(),
@@ -86,95 +84,7 @@ const Login = (userModel, teamModel, cartModel, cartItemModel) => async (req, re
 
     log.info(`user ${user.username} logged`);
 
-    /********************/
-    /** Captive portal **/
-    /********************/
-    let captivePortalSuccess = false;
-    const ip = getIP(req);
-    const ipParts = ip.split('.').map((e) => parseInt(e));
-    const isInUnconnectedNetwork = (
-      ipParts[0] === 172 &&
-      ipParts[1] === 16 &&
-      (ipParts[2] >= 188 && ipParts[2] <= 191) &&
-      (ipParts[3] >= 0 && ipParts[3] <= 255)
-    );
-
-    const isValidPlayer = user.team && user.team.tournamentId !== 5 && user.place;
-    const isOrga = hasOrgaPermission(user.permissions);
-    const streamPermission = user.permissions === 'stream' ? '_stream' : '';
-
-    if (isInUnconnectedNetwork) {
-      if (isValidPlayer || isOrga) {
-        await new Promise((resolve) => {
-          // Replace dots in IP address by dashes
-          const userRedisIP = ip.replace(/\./g, '-');
-
-          // Get MAC of the user from its IP
-          redis.get(userRedisIP, async (err, mac) => {
-            if (err || !mac) {
-              const error = !mac ? `no mac associated to ${userRedisIP} in redis` : JSON.stringify(err);
-              log.error(`captive portal error : ${error}`);
-              // Just quit this function and don't throw an error
-              resolve();
-              return;
-            }
-
-            let network;
-            if(isValidPlayer) {
-              switch(user.team.tournamentId) {
-                case 1:
-                case 2:
-                  network = 'lol';
-                  break;
-                case 3:
-                  network = 'fortnite';
-                  break;
-                case 4:
-                  network = 'csgo';
-                  break;
-                case 6:
-                  network = 'osu';
-                  break;
-                case 7:
-                  network = 'libre';
-                  break;
-              }
-              network += streamPermission;
-            }
-            else if(isOrga) {
-              network = 'staff';
-            }
-
-            if(!network) {
-              return;
-            }
-
-            await new Promise((resolve) => {
-              // Associate to MAC: { network, firstname, lastname, username, email, place }
-              redis.set(
-                mac,
-                JSON.stringify({
-                  network,
-                  firstname: user.firstname,
-                  lastname: user.lastname,
-                  username: user.username,
-                  email: user.email,
-                  place: user.place,
-                }),
-                resolve,
-              );
-            });
-
-            captivePortalSuccess = true;
-            log.info(`captive portal successful for ${user.username}, mac = ${mac}`);
-            resolve();
-          });
-        });
-      }
-      else {
-        log.warn('captive portal refused : not a valid player or orga');
-      }
-    }
+    const captivePortalSuccess = await captivePortal(req, user);
 
     return res
       .status(200)
